requirements.txt
apiclient==1.0.2
DateTime==4.0.1
google-api-python-client==1.4.0
httplib2==0.9.1
oauth2client==1.4.11
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-dateutil==2.4.2
python-redmine==1.1.2
pytz==2015.4
requests==2.7.0
rsa==3.1.4
simplejson==3.7.3
six==1.9.0
uritemplate==0.6
zope.interface==4.1.2
